Move graphQL query to config file
