+++ src/utils/graphqlQueries.js
@@ -27,36 +27,6 @@ export const GET_CASE_DETAIL_DATA_QUERY = gql`
     }
       }`;
 
-export const GET_MY_CASES_DATA_QUERY = gql`
-query subjectsInList($subject_ids: [String!]!) {
-
-  subjectsInList(subject_ids: $subject_ids) {
-    subject_id
-    program
-    study_acronym
-    diagnosis
-    recurrence_score
-    tumor_size
-    er_status
-    pr_status
-    age_at_index
-    survival_time
-    survival_time_unit
-}
-filesOfSubjects(subject_ids: $subject_ids) {
-   subject_id
-    file_description
-    file_format
-    file_name
-    file_size
-    file_type
-    association
-    file_id
-    md5sum
-}
-
-}`;
-
 export const TRIALS_QUERY = gql`{
    clinicalTrials{
     clinical_trial_id
